:beetle: Fix Windows build

+++ src/Plugin.cpp
@@ -106,34 +106,29 @@ void Plugin::init() {
 
   mGuiManager->getGui()->registerCallback("anchorLabels.setIgnoreOverlapThreshold",
       "Higher values will prevent anchor labels to be hidden when they overlap a little.",
-      std::function([this](double value) {
-        mPluginSettings->mIgnoreOverlapThreshold = static_cast<float>(value);
-      }));
-  mPluginSettings->mIgnoreOverlapThreshold.connectAndTouch([this](float value) {
+      std::function([this](double value) { mPluginSettings->mIgnoreOverlapThreshold = value; }));
+  mPluginSettings->mIgnoreOverlapThreshold.connectAndTouch([this](double value) {
     mGuiManager->setSliderValue("anchorLabels.setIgnoreOverlapThreshold", value);
   });
 
   mGuiManager->getGui()->registerCallback("anchorLabels.setScale",
-      "Sets a global scale multiplier for all anchor labels.", std::function([this](double value) {
-        mPluginSettings->mLabelScale = static_cast<float>(value);
-      }));
+      "Sets a global scale multiplier for all anchor labels.",
+      std::function([this](double value) { mPluginSettings->mLabelScale = value; }));
   mPluginSettings->mLabelScale.connectAndTouch(
-      [this](float value) { mGuiManager->setSliderValue("anchorLabels.setScale", value); });
+      [this](double value) { mGuiManager->setSliderValue("anchorLabels.setScale", value); });
 
   mGuiManager->getGui()->registerCallback("anchorLabels.setDepthScale",
       "Higher values will make the scale of the anchor labels depend on their distance to the "
       "observer.",
-      std::function(
-          [this](double value) { mPluginSettings->mDepthScale = static_cast<float>(value); }));
+      std::function([this](double value) { mPluginSettings->mDepthScale = value; }));
   mPluginSettings->mDepthScale.connectAndTouch(
-      [this](float value) { mGuiManager->setSliderValue("anchorLabels.setDepthScale", value); });
+      [this](double value) { mGuiManager->setSliderValue("anchorLabels.setDepthScale", value); });
 
   mGuiManager->getGui()->registerCallback("anchorLabels.setOffset",
       "Specifies the distance between planet and anchor labels.",
-      std::function(
-          [this](double value) { mPluginSettings->mLabelOffset = static_cast<float>(value); }));
+      std::function([this](double value) { mPluginSettings->mLabelOffset = value; }));
   mPluginSettings->mLabelOffset.connectAndTouch(
-      [this](float value) { mGuiManager->setSliderValue("anchorLabels.setOffset", value); });
+      [this](double value) { mGuiManager->setSliderValue("anchorLabels.setOffset", value); });
 
   onLoad();
